Update home-block1-intro.md
Amends designed to (1) spell out the tech a bit more commercially (2) get everyone in blockchain/P2P mode and (3) get the value proposition front and centre. 

Also DG had a suggestion re: branding - Thel/DeCerver are components but should be subsumed within the Eris brand to a certain extent so that Eris is brand-identifiable to the key Series A constituents. Suggest referring to the system as a whole either as "Eris 1.0" or "Eris Stack" for ease of reference.

N.b. have borrowed some language from the PD website - a useful guidepost. Good to remember we have been in full-on Dev Mode for 8 months but we do need to make sure we express our thoughts in English where we can - it'll show we can market.